+++ fd-vscode/webview/wasm/fd_wasm_bg.wasm.d.ts
@@ -3,6 +3,7 @@
 export const memory: WebAssembly.Memory;
 export const __wbg_fdcanvas_free: (a: number, b: number) => void;
 export const fdcanvas_add_animation_to_node: (a: number, b: number, c: number, d: number, e: number, f: number, g: number) => number;
+export const fdcanvas_cancel_drag: (a: number) => number;
 export const fdcanvas_clear_pressed: (a: number) => void;
 export const fdcanvas_compute_guides_for_rect: (a: number, b: number, c: number, d: number, e: number) => [number, number];
 export const fdcanvas_create_edge: (a: number, b: number, c: number, d: number, e: number) => [number, number];